+++ kalamine/layout.py
@@ -1,33 +1,12 @@
 import copy
-import datetime
-import pkgutil
-import re
 import sys
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Optional, Set, Type, TypeVar
-from xml.etree import ElementTree as ET
 
-import click
 import tomli
 import yaml
 
-from .template import (
-    ahk_keymap,
-    ahk_shortcuts,
-    c_deadkeys,
-    c_dk_index,
-    c_keymap,
-    klc_deadkeys,
-    klc_dk_index,
-    klc_keymap,
-    osx_actions,
-    osx_keymap,
-    osx_terminators,
-    web_deadkeys,
-    web_keymap,
-    xkb_keymap,
-)
 from .utils import (
     DEAD_KEYS,
     LAYER_KEYS,
@@ -36,86 +15,12 @@ from .utils import (
     lines_to_text,
     load_data,
     text_to_lines,
+    upper_key,
 )
 
 ###
 # Helpers
 #
-
-
-def get_langid(locale: str) -> str:
-    locale_codes = load_data("win_locales")
-    if locale not in locale_codes:
-        raise ValueError(f"`{locale}` is not a valid locale")
-    return locale_codes[locale]
-
-
-def upper_key(letter: Optional[str], blank_if_obvious: bool = True) -> str:
-    """This is used for presentation purposes: in a key, the upper character
-    becomes blank if it's an obvious uppercase version of the base character."""
-
-    if letter is None:
-        return " "
-
-    custom_alpha = {
-        "\u00df": "\u1e9e",  # ß ẞ
-        "\u007c": "\u00a6",  # | ¦
-        "\u003c": "\u2264",  # < ≤
-        "\u003e": "\u2265",  # > ≥
-        "\u2020": "\u2021",  # † ‡
-        "\u2190": "\u21d0",  # ← ⇐
-        "\u2191": "\u21d1",  # ↑ ⇑
-        "\u2192": "\u21d2",  # → ⇒
-        "\u2193": "\u21d3",  # ↓ ⇓
-        "\u00b5": " ",  # µ (to avoid getting `Μ` as uppercase)
-    }
-    if letter in custom_alpha:
-        return custom_alpha[letter]
-
-    if len(letter) == 1 and letter.upper() != letter.lower():
-        return letter.upper()
-
-    # dead key or non-letter character
-    return " " if blank_if_obvious else letter
-
-
-def substitute_lines(text: str, variable: str, lines: List[str]) -> str:
-    prefix = "KALAMINE::"
-    exp = re.compile(".*" + prefix + variable + ".*")
-
-    indent = ""
-    for line in text.split("\n"):
-        m = exp.match(line)
-        if m:
-            indent = m.group().split(prefix)[0]
-            break
-
-    return exp.sub(lines_to_text(lines, indent), text)
-
-
-def substitute_token(text: str, token: str, value: str) -> str:
-    exp = re.compile("\\$\\{" + token + "(=[^\\}]*){0,1}\\}")
-    return exp.sub(value, text)
-
-
-def load_tpl(layout: "KeyboardLayout", ext: str, tpl: str = "base") -> str:
-    date = datetime.date.today().isoformat()
-    if tpl == "base":
-        if layout.has_altgr or ext.startswith(".RC"):
-            tpl = "full"
-            if layout.has_1dk and ext.startswith(".xkb"):
-                tpl = "full_1dk"
-    bin = pkgutil.get_data(__package__, f"tpl/{tpl}{ext}")
-    if bin is None:
-        return ""
-    out = bin.decode("utf-8")
-    out = substitute_lines(out, "GEOMETRY_base", layout.base)
-    out = substitute_lines(out, "GEOMETRY_full", layout.full)
-    out = substitute_lines(out, "GEOMETRY_altgr", layout.altgr)
-    out = substitute_token(out, "KALAMINE", f"Generated by kalamine on {date}")
-    for key, value in layout.meta.items():
-        out = substitute_token(out, key, value)
-    return out
 
 
 def load_layout(layout_path: Path) -> Dict:
@@ -141,8 +46,8 @@ def load_layout(layout_path: Path) -> Dict:
         return cfg
 
     except Exception as exc:
-        click.echo("File could not be parsed.", err=True)
-        click.echo(f"Error: {exc}.", err=True)
+        print("File could not be parsed.", err=True)
+        print(f"Error: {exc}.", err=True)
         sys.exit(1)
 
 
@@ -266,7 +171,7 @@ class KeyboardLayout:
                 # should bevome ['ab05', 'lsgt', 'ab01', 'ab02', 'ab03', 'ab04']
                 last_row.keys[:6] = [last_row.keys[5]] + last_row.keys[:5]
             else:
-                click.echo(
+                print(
                     "Warning: geometry does not support angle-mod; ignoring the --angle-mod argument"
                 )
 
@@ -495,156 +400,3 @@ class KeyboardLayout:
     def altgr(self) -> List[str]:
         """AltGr layer only."""
         return self._get_geometry([Layer.ALTGR])
-
-    ###
-    # OS-specific drivers: keylayout, klc, xkb, xkb_patch
-    #
-
-    @property
-    def keylayout(self) -> str:
-        """macOS driver"""
-        out = load_tpl(self, ".keylayout")
-        for i, layer in enumerate(osx_keymap(self)):
-            out = substitute_lines(out, "LAYER_" + str(i), layer)
-        out = substitute_lines(out, "ACTIONS", osx_actions(self))
-        out = substitute_lines(out, "TERMINATORS", osx_terminators(self))
-        return out
-
-    @property
-    def ahk(self) -> str:
-        """Windows AHK driver"""
-        out = load_tpl(self, ".ahk")
-        out = substitute_lines(out, "LAYOUT", ahk_keymap(self))
-        out = substitute_lines(out, "ALTGR", ahk_keymap(self, True))
-        out = substitute_lines(out, "SHORTCUTS", ahk_shortcuts(self))
-        return out
-
-    @property
-    def klc(self) -> str:
-        """Windows driver (warning: requires CR/LF + UTF16LE encoding)"""
-        if len(self.meta["name8"]) > 8:
-            raise ValueError("`name8` max length is 8 charaters")
-
-        # check `version` format
-        # it must be `a.b.c[.d]`
-        version = re.compile(r"^\d+\.\d+\.\d+(\.\d+)?$")
-        if version.match(self.meta["version"]) is None:
-            raise ValueError("`version` must be in `a.b.c[.d]` form")
-        locale = self.meta["locale"]
-        langid = get_langid(locale)
-        out = load_tpl(self, ".klc")
-        out = substitute_lines(out, "LAYOUT", klc_keymap(self))
-        out = substitute_lines(out, "DEAD_KEYS", klc_deadkeys(self))
-        out = substitute_lines(out, "DEAD_KEY_INDEX", klc_dk_index(self))
-        out = substitute_token(out, "localeid", f"0000{langid}")
-        out = substitute_token(out, "locale", locale)
-        out = substitute_token(out, "encoding", "utf-16le")
-        return out
-
-    @property
-    def klc_rc(self) -> str:
-        """Windows resource file for C drivers"""
-        out = load_tpl(self, ".RC")
-        # version numbers are in "a,b,c,d" format
-        version = self.meta["version"].replace(".", ",")
-        out = substitute_token(out, "rc_version", version)
-        return out
-
-    @property
-    def klc_c(self) -> str:
-        """Windows keymap file for C drivers"""
-        out = load_tpl(self, ".C")
-        out = substitute_lines(out, "LAYOUT", c_keymap(self))
-        out = substitute_lines(out, "DEAD_KEYS", c_deadkeys(self))
-        out = substitute_lines(out, "DEAD_KEY_INDEX", c_dk_index(self))
-        return out
-
-    @property
-    def xkb_keymap(self) -> str:  # will not work with Wayland
-        """GNU/Linux driver (standalone / user-space)"""
-        out = load_tpl(self, ".xkb_keymap")
-        out = substitute_lines(out, "LAYOUT", xkb_keymap(self, xkbcomp=True))
-        return out
-
-    @property
-    def xkb_symbols(self) -> str:
-        """GNU/Linux driver (xkb patch, system or user-space)"""
-        out = load_tpl(self, ".xkb_symbols")
-        out = substitute_lines(out, "LAYOUT", xkb_keymap(self, xkbcomp=False))
-        return out
-
-    ###
-    # JSON output: keymap (base+altgr layers) and dead keys
-    #
-
-    @property
-    def json(self) -> Dict:
-        """JSON layout descriptor"""
-        return {
-            "name": self.meta["name"],
-            "description": self.meta["description"],
-            "geometry": self.meta["geometry"].lower(),
-            "keymap": web_keymap(self),
-            "deadkeys": web_deadkeys(self),
-            "altgr": self.has_altgr,
-        }
-
-    ###
-    # SVG output
-    #
-
-    @property
-    def svg(self) -> ET.ElementTree:
-        """SVG drawing"""
-
-        svg_ns = "http://www.w3.org/2000/svg"
-        ET.register_namespace("", svg_ns)
-        ns = {"": svg_ns}
-
-        # Parse the SVG template
-        res = pkgutil.get_data(__package__, "tpl/x-keyboard.svg")
-        if res is None:
-            return ET.ElementTree()
-        svg = ET.ElementTree(ET.fromstring(res.decode("utf-8")))
-
-        def set_key_label(key: ET.Element, lvl: int, char: str) -> None:
-            for label in key.findall(f'g/text[@class="level{lvl}"]', ns):
-                if char not in deadkeys:
-                    label.text = char
-                else:  # only show last char for deadkeys
-                    if char == ODK_ID:
-                        label.text = "★"
-                    elif char == "*µ":
-                        label.text = "α"
-                    else:
-                        label.text = char[-1]
-                    label.set("class", f"{label.get('class')} deadKey")
-
-        # Get layout data
-        keymap = web_keymap(self)
-        deadkeys = web_deadkeys(self)
-
-        # Fill-in with layout
-        for name, chars in keymap.items():
-            for key in svg.findall(f'.//g[@id="{name}"]', ns):
-
-                # Print 1-4 level chars
-                for level_num, char in enumerate(chars, start=1):
-                    if level_num == 1:
-                        if chars[1] == upper_key(chars[0], blank_if_obvious=False):
-                            continue
-                    if level_num == 4:
-                        if chars[3] == upper_key(chars[2], blank_if_obvious=False):
-                            continue
-                    set_key_label(key, level_num, char)
-
-                # Print 5-6 levels (1dk deadkeys)
-                if deadkeys and (odk := deadkeys.get(ODK_ID)):
-                    level5 = odk.get(chars[0])
-                    level6 = odk.get(chars[1])
-                    if level5:
-                        set_key_label(key, 5, level5)
-                    if level6 and level6 != upper_key(level5, blank_if_obvious=False):
-                        set_key_label(key, 6, level6)
-
-        return svg
