+++ kalamine/templates/base.xkb_keymap
