+++ kalamine/templates/base.xkb_symbols
