+++ kalamine/templates/full.RC
