+++ kalamine/templates/full.keylayout
