+++ kalamine/templates/full.klc
