+++ kalamine/templates/full.xkb_keymap
