+++ kalamine/templates/full.xkb_symbols
