+++ kalamine/templates/full_1dk.xkb_keymap
